algo-ui/backend/requirements.txt
fastapi
uvicorn[standard]
pandas
python-multipart
numpy
pytz
python-dateutil
# Your trading SDK
growwapi
